Use `next_element` instead of `next` in nokogiri.
For some reason, in testing and development, nokogiri worked just fine,
but in production (heroku) it didn't. Debugging the problem I saw in
development calling `next` from a `tr` nokogiri node returns a `\n\r` as
the next element and not the next `tr` as expected. So I needed to add
another `next` to rich the `tr` a wanted to. (`node.next.next`). Changing
`next` to `next_element` fixed it in development and could work in heroku.